+++ R/projects.R
@@ -121,14 +121,10 @@ setMethod("entity", "CrunchProject", function (x) {
 
 #' @rdname delete
 #' @export
-setMethod("delete", "CrunchProject", function (x, confirm=requireConsent(), ...) {
-    if (!missing(confirm)) {
-        warning("The 'confirm' argument is deprecated. See ?with_consent.",
-            call.=FALSE)
-    }
+setMethod("delete", "CrunchProject", function (x, ...) {
     prompt <- paste0("Really delete project ", dQuote(name(x)), "? ",
         "This cannot be undone.")
-    if (confirm && !askForPermission(prompt)) {
+    if (!askForPermission(prompt)) {
         halt("Must confirm deleting project")
     }
     u <- self(x)