+++ R/tab-book.R
@@ -128,18 +128,19 @@ setMethod("initialize", "MultitableResult", function (.Object, ...) {
     .Object <- callNextMethod(.Object, ...)
     ## The first cube in the results list is the "total" column. It's a 1-D
     ## cube, add a second "dimension" so that it appears to be 2-D, n x 1
-    .Object$result[[1]]$result$dimensions[[2]] <- list(
-        type=list(
-            class="enum",
-            elements=list(
-                list(id=0, value="", missing=FALSE)
+    .Object$result[[1]]$result$dimensions <- c(.Object$result[[1]]$result$dimensions,
+        list(list(
+            type=list(
+                class="enum",
+                elements=list(
+                    list(id=0, value="", missing=FALSE)
+                )
+            ),
+            references=list(
+                alias="total",
+                name="Total"
             )
-        ),
-        references=list(
-            alias="total",
-            name="Total"
-        )
-    )
+        )))
     .Object$result <- lapply(.Object$result, CrunchCube)
     return(.Object)
 })