+++ tests/testthat/test-copy-variable.R
@@ -16,7 +16,7 @@ with_mock_HTTP({
                 show_codes=FALSE,
                 column_width=NULL
             ),
-            expr=list(
+            derivation=list(
                 `function`="copy_variable",
                 args=list(
                     list(variable="api/datasets/1/variables/gender/")