+++ tests/testthat/test-dataset-entity.R
@@ -253,10 +253,6 @@ with_mock_HTTP({
     })
     test_that("Dataset deleting", {
         expect_error(delete(ds), "Must confirm") ## New non-interactive behavior
-        expect_warning(
-            expect_error(delete(ds, confirm=TRUE), "Must confirm"),
-            "The 'confirm' argument is deprecated."
-        )
         with_consent(expect_DELETE(delete(ds), self(ds)))  ## No warning
     })
 })
@@ -363,9 +359,9 @@ with_test_authentication({
         test_that("Setting default weight", {
             settings(ds)$weight <- self(ds$name)
             expect_identical(settings(ds)$weight, self(ds$name))
-            ## And it should now be in the weight_variables catalog too right?
-            expect_identical(urls(ShojiCatalog(crGET(shojiURL(ds, "catalogs",
-                "weight_variables")))), self(ds$name))
+            ## And it should now be in the weight variables order too
+            expect_identical(urls(ShojiOrder(crGET(shojiURL(variables(ds),
+                "orders", "weights")))), self(ds$name))
             ## Can also remove the setting
             settings(ds)$weight <- NULL
             expect_null(settings(ds)$weight)