+++ tests/testthat/test-weighting.R
@@ -111,4 +111,5 @@ with_test_authentication({
                 array(c(110, 100), dim=2L, dimnames=list(v4=c("B", "C"))))
         })
     })
+
 })